+++ samples/Digipolis.Web.SampleApi/Controllers/ValuesController.cs
@@ -67,7 +67,7 @@ namespace Digipolis.Web.SampleApi.Controllers
             int total;
             var values = _valueLogic.GetAll(criteria, out total);
             //var result = queryOptions.ToPagedResult(values, total, "kevin", new { test = 0 });
-            var result = criteria.ToPagedResult<ValueDto, EmbeddedValueDto>(values, total, "Get", "Values", new { test = 0 });
+            var result = criteria.ToPagedResult<ValueDto, EmbeddedValueDto>(values, total);
             return Ok(result);
         }
         /// <summary>